Mopho 2: Came to a better understanding of what the ValueTree::getChild() functions do. They do not return a reference to the child tree, they create a distinct ValueTree object which mirrors the source child tree. Changes to the source tree are reflected in the new tree, but the new tree has a life span separate from the source tree. Therefore, functions like ExposedParamsRandomizationOptions::addListenerToChildTreeForParam() will not work. The ValueTree created by getChild() only exists inside the function and adding a listener to it is useless. For a class to listen to a child of a ValueTree held in another class, it is necessary to declare a member ValueTree in the listener class, initialize it using getChild() to make it a mirror of the source child tree, then add the listener to the member tree. In ExposedParamsRandomizationOptions, I eliminated the addListener and removeListener functions and declared and defined getChildTreeForParam() and getChildTreeForSeqTrack(). In AllowRepeatChoicesToggle, AllowRepeatChoicesToggle_SeqTrackStep, GUI_Layer_AllowedChoices_SeqTrack, and KnobForSeqTrackProbability, I added member ValueTrees, made them mirror the relevant child trees in ExposedParamsRandomizationOptions, and added/removed listeners to these member trees. In VoicesBanks, eliminated addListenerToCustomVoiceNameStringsTree() and removeListenerFromCustomVoiceNameStringsTree(), and declared and defined getVoiceNameStringsChildTreeForCustomBank(). In VoiceSlots, used the mirror tree method to listen to changes to the voice name trees for the custom banks.

+++ MophoGUI/MophoGUI 2.0/Source/randomization/rndm_1_comp_AllowChoiceToggles_VoiceNameChar.cpp
@@ -61,7 +61,7 @@ const bool AllowChoiceToggles_VoiceNameChar::noChoiceIsAllowed() {
 }
 
 void AllowChoiceToggles_VoiceNameChar::restoreToggles() {
-	for (auto choiceNum = 0; choiceNum < numberOfChoices; ++choiceNum) {
+	for (auto choiceNum = (uint8)0; choiceNum < numberOfChoices; ++choiceNum) {
 		auto isAllowed{ randomization->choiceIsAllowedForVoiceNameCharParam(choiceNum, paramIndex) };
 		allowedChoiceToggles[choiceNum]->setToggleState(isAllowed ? true : false, dontSendNotification);
 	}
